app/pyproject.toml
[tool.poetry]
name = "aws-demo"
version = "0.1.0"
description = ""
authors = ["SudeepRed <sudeep19313@iiitd.ac.in>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.3"
mysql-connector-python = "^8.0.32"
pytest-playwright = "^0.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
